# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WeDo Backend API",
    "version": "1.0.0",
    "description": "OpenAPI specification for the WeDo backend."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": { "type": "string" },
          "message": { "type": "string" }
        },
        "required": ["code", "message"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        },
        "required": ["error"]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "type": { "type": "string" },
          "title": { "type": "string" },
          "message": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "read": { "type": "boolean" },
          "activityId": { "type": "string", "nullable": true },
          "requestUserId": { "type": "string", "nullable": true },
          "actionRequired": { "type": "boolean" },
          "activityTitle": { "type": "string", "nullable": true }
        },
        "required": ["id", "type", "title", "message", "timestamp", "read", "actionRequired"]
      },
      "NotificationsResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Notification" } },
          "nextCursor": { "type": "string", "nullable": true },
          "hasMore": { "type": "boolean" }
        },
        "required": ["items", "hasMore"]
      },
      "UpdateNotificationRequest": {
        "type": "object",
        "properties": {
          "read": { "type": "boolean" }
        }
      },
...
      "CreateSubscriptionRequest": {
        "type": "object",
        "properties": {
          "userId": { "type": "integer" }
        },
        "required": ["userId"]
      },
      "CreateSubscriptionResponse": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "subscribedAt": { "type": "string", "format": "date-time" },
          "isPinned": { "type": "boolean" }
        },
        "required": ["userId", "subscribedAt", "isPinned"]
      },
      "UpdateSubscriptionRequest": {
        "type": "object",
        "properties": {
          "isPinned": { "type": "boolean" }
        }
      },
      "SubscriptionDeleteResponse": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "deleted": { "type": "boolean" }
        },
        "required": ["userId", "deleted"]
      },
      "SubscriptionUpdateResponse": {
        "type": "object",
        "properties": {
          "userId": { "type": "string" },
          "isPinned": { "type": "boolean" }
        },
        "required": ["userId", "isPinned"]
      },
...
    }
  },
  "security": [
    { "BearerAuth": [] }
  ],
  "paths": {
...
    "/me/device-token": {
      "post": {
        "tags": ["notifications"],
        "summary": "Register FCM device token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" }
                },
                "required": ["token"]
              }
            }
          }
        },
        "responses": {
          "204": { "description": "Token saved successfully" },
          "400": {
            "description": "Token missing",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/me/notifications": {
      "get": {
        "tags": ["notifications"],
        "summary": "List current user notifications",
        "parameters": [
          { "name": "unreadOnly", "in": "query", "schema": { "type": "boolean" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Notifications page",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/NotificationsResponse" }
              }
            }
          }
        }
      }
    },
    "/notifications/{notification_id}": {
      "patch": {
        "tags": ["notifications"],
        "summary": "Update notification read state",
        "parameters": [
          { "name": "notification_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateNotificationRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated notification state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Notification" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["notifications"],
        "summary": "Delete notification",
        "parameters": [
          { "name": "notification_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": {
            "description": "Notification deleted"
          }
        }
      }
    },
...
    "/subscriptions": {
      "post": {
        "tags": ["subscriptions"],
        "summary": "Create subscription",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateSubscriptionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Subscription created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateSubscriptionResponse" }
              }
            }
          }
        }
      }
    },
    "/subscriptions/{user_id}": {
      "patch": {
        "tags": ["subscriptions"],
        "summary": "Update subscription",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateSubscriptionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated subscription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubscriptionUpdateResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["subscriptions"],
        "summary": "Delete subscription",
        "parameters": [
          { "name": "user_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Deleted subscription",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubscriptionDeleteResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}